% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decomposition_curve.R
\name{Decomposition_curve}
\alias{Decomposition_curve}
\title{Decomposition_curve}
\usage{
Decomposition_curve(exp_index, legendre_order, data, alpha = 0)
}
\arguments{
\item{exp_index}{vector of the time point}

\item{legendre_order}{scalar of LOP order}

\item{data}{Net effect value of the mean curve(matrix of observed data)}

\item{alpha}{Adjust the degree of punishment. The default value is 0. If necessary, the reasonable range is between 0 and 1}
}
\value{
Type is list.The list contains LOP parameter、interactions between sites、the elements that make up a network.
}
\description{
Decomposition_curve
}
