% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fourteen_cluster_mean_effect}
\alias{fourteen_cluster_mean_effect}
\title{Mean curves of 14 types of modules}
\format{
A data frame of mean curves of 14 types of modules
}
\usage{
data(fourteen_cluster_mean_effect)
}
\description{
Mean curves of 14 types of modules
}
\keyword{datasets}
