% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fourth_effect_cluster}
\alias{fourth_effect_cluster}
\title{The optimal cluster number is 14}
\format{
According to BIC, the optimal cluster number is 14
}
\usage{
data(fourth_effect_cluster)
}
\description{
The optimal cluster number is 14
}
\keyword{datasets}
