% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funcluster.R
\name{Funcluster}
\alias{Funcluster}
\title{Funcluster}
\usage{
Funcluster(data, times, rep, k, legendre_order)
}
\arguments{
\item{data}{The net genetic effects of all loci were calculated by Get_effect function(remarker_effect).}

\item{times}{rep A series of times for measuring phenotypic data}

\item{rep}{Sets the number of funclusters specified, usually 1}

\item{k}{Specify the number of clusters.}

\item{legendre_order}{Specifies the order of Legendre polynomials}
}
\value{
Type is list.The list contains BIC Maximum、likelihood、Legendre polynomials parameters、Mean curve parameters for each cluster.
}
\description{
According to modularity theory,functional clustering (FunClu) was
introduced to cluster genetic effects into different functional modules.
A hybrid of the EM and simplex algorithms were implanted to obtain the functional modules.
}
