% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funmap.R
\name{Funmap_function}
\alias{Funmap_function}
\title{FunMap}
\usage{
Funmap_function(
  funmap_data,
  funmap_data_snp,
  times,
  rownumber = dim(funmap_data_snp)[1]
)
}
\arguments{
\item{funmap_data}{The phenotypic data.Remove or replace missing values in the data; Log-transformed the phenotypic data by command log() if the number in the data vary widely is recommended, such as numbers differ by more than five order of magnitude.}

\item{funmap_data_snp}{Genotypic data.only accept numeric value for calculation, SNP genotype data should be converted into 0,1 matrix.}

\item{times}{A series of times for measuring phenotypic data}

\item{rownumber}{The default value is the total amount of genotype data.For simple tests, enter any total value between 1- total value of genotype data.}
}
\value{
Type is list.The list contains H0 result、H1 result、All H1 value、LR of each SNP.
}
\description{
FunMap is crucial to this model, for it excavates how specific QTLs
determines the complex trait. The mean vector and covariance
structure should be modeled according to the design of the experiment.
}
