% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genetic_effect}
\alias{genetic_effect}
\title{genetic effect data}
\format{
A data frame of 28 columns and 8305 rows
}
\usage{
genetic_effect
}
\description{
A dataset containing calculated generic effect from geno and pheno in this dataset
}
\keyword{datasets}
