% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geno}
\alias{geno}
\title{genotypic data}
\format{
A data frame with 8305 rows and 108 variables, contain row names as SNP ID:
}
\usage{
data(geno)
}
\description{
A dataset containing genotypic data of 106 Populus euphratica samples
}
\keyword{datasets}
