% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plot.R
\name{get_after}
\alias{get_after}
\title{convert ODE results to basic network plot table}
\usage{
get_after(i)
}
\arguments{
\item{i}{list result from get_all_net(only use a single list)}
}
\value{
a table with basic information to plot network
}
\description{
convert ODE results to basic network plot table
}
