% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODE_solving.R
\name{get_all_net}
\alias{get_all_net}
\title{convert all ODE result into a list(use when multiply network needed)}
\usage{
get_all_net(input)
}
\arguments{
\item{input}{list contain the ODE solving result from get_ode_par function}
}
\value{
list of output for all data
}
\description{
convert all ODE result into a list(use when multiply network needed)
}
