% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biFunMap.R
\name{get_biFunMap_result}
\alias{get_biFunMap_result}
\title{perform biFunMap for all SNP in parallel mode}
\usage{
get_biFunMap_result(geno_df, pheno_df, times, write_data = TRUE)
}
\arguments{
\item{geno_df}{dataframe of Genotypic data}

\item{pheno_df}{dataframe of Phenotypic data}

\item{times}{vector for the time point}

\item{write_data}{write result table for further analysis}
}
\value{
a list of LR values
}
\description{
perform biFunMap for all SNP in parallel mode
}
