% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plot.R
\name{get_colour}
\alias{get_colour}
\title{generate color for edge}
\usage{
get_colour(i, color_data)
}
\arguments{
\item{i}{row number}

\item{color_data}{dataframe of all color for edge}
}
\value{
Hex code
}
\description{
generate color for edge
}
