% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODE_solving.R
\name{get_curve_data}
\alias{get_curve_data}
\title{helper function of plot decomposition_curve}
\usage{
get_curve_data(input1, input2, i)
}
\arguments{
\item{input1}{list object from get_ode_par function}

\item{input2}{list object from get_all_net function}

\item{i}{scalar of row number}
}
\value{
datafrmae for plot
}
\description{
helper function of plot decomposition_curve
}
