% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODE_solving.R
\name{get_decomposition_plot}
\alias{get_decomposition_plot}
\title{plot decomposition_curve}
\usage{
get_decomposition_plot(input1, input2, i)
}
\arguments{
\item{input1}{list object from get_ode_par function}

\item{input2}{list object from get_all_net function}

\item{i}{scalar of i row number to plot}
}
\value{
plot
}
\description{
plot decomposition_curve
}
