% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_effect.R
\name{Get_effect}
\alias{Get_effect}
\title{Get_effect}
\usage{
Get_effect(funmap_data_snp, funmap_data_snp_par, times)
}
\arguments{
\item{funmap_data_snp}{Genotypic data.only accept numeric value for calculation, SNP genotype data should be converted into 0,1 matrix.}

\item{funmap_data_snp_par}{The H1 parameters of each gene locus were calculated by FunMap function.code：}

\item{times}{A series of times for measuring phenotypic data.}
}
\value{
The net genetic effect of each gene locus (SNP).Type is data frame.
}
\description{
The net genetic effect of each gene locus (SNP) was calculated
}
