% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plot.R
\name{get_effect_type}
\alias{get_effect_type}
\title{convert >=0 value to '+' and < 0 value = '-'}
\usage{
get_effect_type(i, links)
}
\arguments{
\item{i}{scalar of row number}

\item{links}{matrix contain links data of network}
}
\value{
a characte 0f '+' or '-'
}
\description{
convert >=0 value to '+' and < 0 value = '-'
}
