% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biFunMap.R
\name{get_generic_effect_plot}
\alias{get_generic_effect_plot}
\title{plot generic effect of random choose number = number}
\usage{
get_generic_effect_plot(generic_effect, number)
}
\arguments{
\item{generic_effect}{dataframe of calculated generic effect data}

\item{number}{scalar of number of SNPs' effect curve want to show(must <= nrow(generic_effect))}
}
\value{
line plot of generic effect
}
\description{
plot generic effect of random choose number = number
}
