% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biFunMap.R
\name{get_genetic_effect_plot}
\alias{get_genetic_effect_plot}
\title{plot genetic effect of random choose number = number}
\usage{
get_genetic_effect_plot(genetic_effect, number)
}
\arguments{
\item{genetic_effect}{dataframe of calculated genetic effect data}

\item{number}{scalar of number of SNPs' effect curve want to show(must <= nrow(genetic_effect))}
}
\value{
line plot of genetic effect
}
\description{
plot genetic effect of random choose number = number
}
