% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biFunClu.R
\name{get_init_par}
\alias{get_init_par}
\title{generate inital parameters for biFunClu}
\usage{
get_init_par(
  data,
  k,
  legendre_order,
  times,
  init_SAD_par = c(1.06, 0.25, 1.15, 0.18)
)
}
\arguments{
\item{data}{dataframe contain numeric values for further analysis}

\item{k}{scalar indicate the number of cluster}

\item{legendre_order}{scalar of legendre polynomials}

\item{times}{vector of time point}

\item{init_SAD_par}{vector of length 4 for biSAD1 covariance matrix}
}
\value{
list contain initial parameters for biFunClu
}
\description{
generate inital parameters for biFunClu
}
