% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biFunMap.R
\name{get_LR_effect}
\alias{get_LR_effect}
\title{perform biFunMap for a SNP}
\usage{
get_LR_effect(
  marker,
  geno_df,
  pheno_df,
  times,
  init_sd_par = c(0.95, 12, 1.02, 8),
  simplify = FALSE
)
}
\arguments{
\item{marker}{scalar indicate which SNP to use(row number)}

\item{geno_df}{dataframe of Genotypic data}

\item{pheno_df}{dataframe of Phenotypic data}

\item{times}{vector for the time point}

\item{init_sd_par}{vector of length four for biSAD1 covariance matrix}

\item{simplify}{default is FALSE, only used in permutation to return less result}
}
\value{
the LR value, generic effect and calculated parameters
}
\description{
perform biFunMap for a SNP
}
