% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biFunMap.R
\name{get_manh_plot}
\alias{get_manh_plot}
\title{manhattan plot}
\usage{
get_manh_plot(geno_df, LR_result, threshold = 20)
}
\arguments{
\item{geno_df}{dataframe of Genotypic data(contain addioitonal information)}

\item{LR_result}{vector of length = nrow(geno_df), contain all LR value}

\item{threshold}{scalar for LR threthold}
}
\value{
manhattan plot
}
\description{
manhattan plot
}
