% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plot.R
\name{get_max_effect}
\alias{get_max_effect}
\title{calculate maximum effect of a network}
\usage{
get_max_effect(k)
}
\arguments{
\item{k}{list result from get_all_net}
}
\value{
vector of effect
}
\description{
calculate maximum effect of a network
}
