% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biFunMap.R
\name{get_mean_curve_plot}
\alias{get_mean_curve_plot}
\title{plot mean curve for biFunMap results}
\usage{
get_mean_curve_plot(pheno_df, times, init_sd_par = c(0.95, 12, 1.02, 8))
}
\arguments{
\item{pheno_df}{dataframe of Phenotypic data}

\item{times}{vector for the time point}

\item{init_sd_par}{vector of length four for biSAD1 covariance matrix}
}
\value{
a plot with mean curve for biFunMap
}
\description{
plot mean curve for biFunMap results
}
