% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODE_solving.R
\name{get_module_data}
\alias{get_module_data}
\title{generate module dataframe}
\usage{
get_module_data(data_par, times)
}
\arguments{
\item{data_par}{dataframe of clustered results contain parameters for each cluster}

\item{times}{vector of the time point}
}
\value{
list contain two dataframe
}
\description{
generate module dataframe
}
