% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plot.R
\name{get_net_output}
\alias{get_net_output}
\title{convert ODE results to basic network plot table for Cytoscape}
\usage{
get_net_output(input1, input2, write_data = TRUE)
}
\arguments{
\item{input1}{list result from get_ode_par}

\item{input2}{list result from get_all_net}

\item{write_data}{write the result, default is TRUE}
}
\value{
a list with node and link attribute
}
\description{
convert ODE results to basic network plot table for Cytoscape
}
