% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODE_solving.R
\name{get_ode_output}
\alias{get_ode_output}
\title{helper function to convert ODE result}
\usage{
get_ode_output(relationship, par, effect, times, order)
}
\arguments{
\item{relationship}{list contain the result of lasso-based variable election}

\item{par}{vector conatin LOP parameters}

\item{effect}{matrix of observed data}

\item{times}{vector of time point}

\item{order}{scalar of LOP order}
}
\value{
list contain bunch of useful output result for a row
}
\description{
helper function to convert ODE result
}
