% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biFunMap.R
\name{get_permutation}
\alias{get_permutation}
\title{permutation test in parallel mode}
\usage{
get_permutation(n = 1000, geno_df, pheno_df, times)
}
\arguments{
\item{n}{scalar the number of times permutation performed, default = 1000}

\item{geno_df}{dataframe of Genotypic data}

\item{pheno_df}{dataframe of Phenotypic data}

\item{times}{vector for the time point}
}
\value{
a list of LR values, each list contain LR results equal to SNP numbers
}
\description{
permutation test in parallel mode
}
