% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biFunMap.R
\name{get_plot_data}
\alias{get_plot_data}
\title{helper function to convert plot data}
\usage{
get_plot_data(data)
}
\arguments{
\item{data}{dataframe of Phenotypic data}
}
\value{
a long dataframe for plot
}
\description{
helper function to convert plot data
}
