% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biFunMap.R
\name{get_subset_data}
\alias{get_subset_data}
\title{select subset data}
\usage{
get_subset_data(data, cluster)
}
\arguments{
\item{data}{dataframe of calculated genetic effect data}

\item{cluster}{scalar of cluster to choose}
}
\value{
genetic effect dataframe only conatin cluster=i
}
\description{
select subset data
}
