% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODE_solving.R
\name{get_value}
\alias{get_value}
\title{estimate parameters for LOP}
\usage{
get_value(effect, relationship, times, order)
}
\arguments{
\item{effect}{matrix of observed data}

\item{relationship}{list contain the result of lasso-based variable election}

\item{times}{vector of time point}

\item{order}{scalar of LOP order}
}
\value{
matrix of estimated LOP parameters
}
\description{
estimate parameters for LOP
}
