% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biFunMap.R
\name{L0}
\alias{L0}
\title{calculate Loglikelihood under h0}
\usage{
L0(y_all, par, times)
}
\arguments{
\item{y_all}{dataframe contain all phenotypic data except missing marker(drop marker=9)}

\item{par}{vector for modified logistic growth curve}

\item{times}{vector for the time point}
}
\value{
the Loglikelihood value
}
\description{
calculate Loglikelihood under h0
}
