% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biFunMap.R
\name{L1}
\alias{L1}
\title{calculate Loglikelihood under h1}
\usage{
L1(pheno_0, pheno_1, par, times)
}
\arguments{
\item{pheno_0}{dataframe contain all phenotypic data with marker=0}

\item{pheno_1}{dataframe contain all phenotypic data with marker=1}

\item{par}{vector for modified logistic growth curve}

\item{times}{vector for the time point}
}
\value{
the Loglikelihood value
}
\description{
calculate Loglikelihood under h1
}
