% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LR}
\alias{LR}
\title{LR(log-likelihood-ratio) data}
\format{
A vector length = 8305
}
\usage{
LR
}
\description{
A dataset containing calculated LR value from geno and pheno in this dataset
}
\keyword{datasets}
