% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plot.R
\name{network_plot}
\alias{network_plot}
\title{generate color for edge}
\usage{
network_plot(k, title, max_effect, color_size = 100, save_plot = TRUE)
}
\arguments{
\item{k}{list result from get_all_net}

\item{title}{text for plot title}

\item{max_effect}{matrix for}

\item{color_size}{scalar control graident size of colour}

\item{save_plot}{save pdf file}
}
\value{
network plot
}
\description{
generate color for edge
}
