% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODE_solving.R
\name{ode_optimize}
\alias{ode_optimize}
\title{calculate least-square for observed and fitted data}
\usage{
ode_optimize(pars, ind, dep, times, data, order, effect)
}
\arguments{
\item{pars}{matrix of LOP parameters for ind and dep growth curve}

\item{ind}{the independent growth curve id}

\item{dep}{the dependent growth curve id}

\item{times}{vector of time point}

\item{data}{dataframe of observed data}

\item{order}{scalar of LOP order}

\item{effect}{matrix of observed data}
}
\value{
scalar of least-square error
}
\description{
calculate least-square for observed and fitted data
}
