% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutaion.R
\name{permutaion}
\alias{permutaion}
\title{Permutaion}
\usage{
permutaion(
  funmap_data,
  funmap_data_snp,
  times,
  permutaion_times,
  rownumber = dim(funmap_data_snp)[1]
)
}
\arguments{
\item{funmap_data}{The phenotypic data.Remove or replace missing values in the data; Log-transformed the phenotypic data by command log() if the number in the data vary widely is recommended, such as numbers differ by more than five order of magnitude.}

\item{funmap_data_snp}{Genotypic data.only accept numeric value for calculation, SNP genotype data should be converted into 0,1 matrix.}

\item{times}{A series of times for measuring phenotypic data}

\item{permutaion_times}{Number of permutation tests.}

\item{rownumber}{The default value is the total amount of genotype data.For simple tests, enter any total value between 1- total value of genotype data.}
}
\value{
LR of each SNP calculated by FunMap 1000 times
}
\description{
By shuffling the relationship between genotype and phenotype repeatedly,
a series of maximum logarithmic likelihood ratios were calculated and critical thresholds were determined from their distribution.
The logarithmic likelihood ratio is compared to the threshold.If the null hypothesis is rejected, it is said to be a significant QTL.
}
