% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pheno}
\alias{pheno}
\title{phenotypic data}
\format{
A data frame with 106 rows and 28 variables, contain row names as sample ID:
}
\usage{
data(pheno)
}
\description{
A dataset containing phenotypic data of 106 Populus euphratica samples
under salt-free and salt-stress environment measured 14 times
}
\keyword{datasets}
