% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manhattan_plot.R
\name{plot_Manhattan_fn}
\alias{plot_Manhattan_fn}
\title{plot_Manhattan}
\usage{
plot_Manhattan_fn(
  funmap_data_snp,
  LR,
  threshold_value,
  x_axis_name = "S. aureus Genome (Mb)"
)
}
\arguments{
\item{funmap_data_snp}{Genotypic data.only accept numeric value for calculation, SNP genotype data should be converted into 0,1 matrix.}

\item{LR}{LR values for each gene locus.}

\item{threshold_value}{The threshold calculated by the Permutaion function.}

\item{x_axis_name}{The name of the X-axis can be customized.}
}
\value{
Manhattan plot
}
\description{
plot_Manhattan
}
