% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{remarker_data_par}
\alias{remarker_data_par}
\title{Parameters of significant loci}
\format{
A data frame of parameters of significant loci,Used in section 3
}
\usage{
data(remarker_data_par)
}
\description{
Parameters of significant loci
}
\keyword{datasets}
