% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{remarker_effect}
\alias{remarker_effect}
\title{The net genetic effects of significant loci}
\format{
A data frame of the net genetic effects of significant loci
}
\usage{
data(remarker_effect)
}
\description{
The net genetic effects of significant loci
}
\keyword{datasets}
