% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plot.R
\name{replace_character}
\alias{replace_character}
\title{replace cut function results to numeric vector}
\usage{
replace_character(i)
}
\arguments{
\item{i}{characte of cut function results}
}
\value{
cleaned numeric vector
}
\description{
replace cut function results to numeric vector
}
