% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s}
\alias{s}
\title{The result calculated by FunMap function}
\format{
A list of the result calculated by FunMap function
}
\usage{
data(s)
}
\description{
The result calculated by FunMap function
}
\keyword{datasets}
