% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{S_mo}
\alias{S_mo}
\title{phenotypic data}
\format{
A data frame with 100 rows and 11 variables
}
\usage{
data(S_mo)
}
\description{
A dataset containing phenotypic data of 100 staphylococcus aureus samples
under a series of time(0.5 h, 1 h, 1.5 h, 2 h, 4 h, 6 h, 8 h, 10 h, 12 h, 16 h, 20 h) were measured
}
\keyword{datasets}
