% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{S_SNP}
\alias{S_SNP}
\title{genotypic data}
\format{
A data frame with 12849 rows and 100 variables
}
\usage{
data(S_SNP)
}
\description{
A dataset containing genotypic data of 100 staphylococcus aureus samples
}
\keyword{datasets}
