% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biFunClu.R
\name{get_cluster_base_plot}
\alias{get_cluster_base_plot}
\title{plot cluster result base plot}
\usage{
get_cluster_base_plot(clustered_data)
}
\arguments{
\item{clustered_data}{clustered data with rownames in first column}
}
\value{
a plot
}
\description{
plot cluster result base plot
}
