% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODE_solving.R
\name{get_effect}
\alias{get_effect}
\title{generate a growth curve(ind or dep effect curve)}
\usage{
get_effect(pars, effect, times, order, y0)
}
\arguments{
\item{pars}{vector of LOP pars}

\item{effect}{scalar of observed genetic or any other data of a time point}

\item{times}{vector of time point}

\item{order}{the order of LOP}

\item{y0}{scalar of initial y}
}
\value{
vector of generated curve
}
\description{
generate a growth curve(ind or dep effect curve)
}
